requirements.txt
annotated-types==0.7.0
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.2
coverage==7.6.10
cycler==0.12.1
distlib==0.3.9
filelock==3.17.0
fire==0.7.0
fonttools==4.57.0
fsspec==2025.3.2
identify==2.6.6
idna==3.10
iniconfig==2.0.0
jinja2==3.1.6
kiwisolver==1.4.8
markupsafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
mypy==1.14.1
mypy-extensions==1.0.0
networkx==3.4.2
nodeenv==1.9.1
numpy==2.1.1
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.6
pluggy==1.5.0
pre-commit==4.1.0
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.10.6
pydantic-core==2.27.2
pyparsing==3.2.3
pytest==8.3.4
python-dateutil==2.9.0.post0
-e file:///D:/Clouddisk/Dropbox/01-Research/2_co-research/HT_Sui/disp_track
pytz==2025.2
pyyaml==6.0.2
requests==2.32.3
ruff==0.9.3
scipy==1.15.2
seaborn==0.13.2
setuptools==79.0.1
six==1.17.0
sympy==1.13.3
termcolor==2.5.0
torch==2.7.0+cu118
torchaudio==2.7.0+cu118
torchvision==0.22.0+cu118
tqdm==4.67.1
typing-extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.116
ultralytics-thop==2.0.14
urllib3==2.4.0
virtualenv==20.29.1
